+++ candle-core/src/dummy_cuda_backend.rs
@@ -146,7 +146,7 @@ impl crate::backend::BackendStorage for CudaStorage {
         _: (usize, usize, usize, usize),
         _: &Layout,
         _: &Layout,
-        _: Option<f64>,
+        _: Option<f32>,
     ) -> Result<Self> {
         Err(Error::NotCompiledWithCudaSupport)
     }
