+++ candle-core/src/dummy_metal_backend.rs
@@ -158,7 +158,7 @@ impl crate::backend::BackendStorage for MetalStorage {
         _: (usize, usize, usize, usize),
         _: &Layout,
         _: &Layout,
-        _: Option<f64>,
+        _: Option<f32>,
     ) -> Result<Self> {
         Err(Error::NotCompiledWithMetalSupport)
     }
